+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockWallBannerBehavior.java
@@ -1,6 +1,8 @@
 package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.component.BlockBlockEntityHolderComponent;
+import org.allaymc.api.blockentity.interfaces.BlockEntityBanner;
 
-public interface BlockWallBannerBehavior extends BlockBehavior {
+public interface BlockWallBannerBehavior extends BlockBehavior, BlockBlockEntityHolderComponent<BlockEntityBanner> {
 }