+++ server/src/main/java/org/allaymc/server/block/component/BlockLoomBaseComponentImpl.java
@@ -1,0 +1,49 @@
+package org.allaymc.server.block.component;
+
+import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.dto.PlayerInteractInfo;
+import org.allaymc.api.block.type.BlockState;
+import org.allaymc.api.block.type.BlockType;
+import org.allaymc.api.container.ContainerTypes;
+import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.math.position.Position3i;
+import org.allaymc.api.world.Dimension;
+import org.joml.Vector3ic;
+
+import static org.allaymc.api.block.property.type.BlockPropertyTypes.DIRECTION_4;
+
+/**
+ * @author daoge_cmd
+ */
+public class BlockLoomBaseComponentImpl extends BlockBaseComponentImpl {
+    public BlockLoomBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
+        super(blockType);
+    }
+
+    @Override
+    public boolean place(Dimension dimension, BlockState blockState, Vector3ic placeBlockPos, PlayerInteractInfo placementInfo) {
+        if (placementInfo != null) {
+            blockState = blockState.setPropertyValue(DIRECTION_4, placementInfo.player().getHorizontalFace().opposite().getHorizontalIndex());
+        }
+        return dimension.setBlockState(placeBlockPos.x(), placeBlockPos.y(), placeBlockPos.z(), blockState, placementInfo);
+    }
+
+    @Override
+    public boolean onInteract(ItemStack itemStack, Dimension dimension, PlayerInteractInfo interactInfo) {
+        if (super.onInteract(itemStack, dimension, interactInfo)) {
+            return true;
+        }
+
+        var player = interactInfo.player();
+        if (player.isSneaking()) {
+            return false;
+        }
+
+        if (player.isActualPlayer()) {
+            var loomContainer = player.getContainer(ContainerTypes.LOOM);
+            loomContainer.setBlockPos(new Position3i(interactInfo.clickedBlockPos(), interactInfo.player().getDimension()));
+            loomContainer.addViewer(player.getController());
+        }
+        return true;
+    }
+}